core: apply graph reduction after build
Build the complete graph, always, independently from the state of the
modules. Annotate each nodes as installed or not dependending on pkg
status. Annotate each edge between nodes as optional if declared as an
optional dependency.

Filter out of the graph the nodes which package missing and their
dependents if that dependent only has optional edges to missing nodes.

It is still the responsiblity of the node declaring a dependency to
manage its behavior depending if the node is missing or not.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>